bot-service/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
databases==0.9.0
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
idna==3.6
ormar==0.9.4
pydantic==2.6.3
pydantic_core==2.16.3
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.27.1
llama-cpp-python==0.2.55
pytest==7.2.2
pytest-cov==4.0.0
ruff==v0.0.260
coverage-badge~=1.1.0
black==23.3.0

